requirements.txt
tensorflow>=2.13.0
numpy>=1.21.0
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
pymysql>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyjwt>=2.8.0
werkzeug>=2.3.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
pyjwt>=2.3.0
python-multipart>=0.0.5
